A basic board class and a main class which constructs one
The Board class currently implements the game world using a 2-D array of a fixed size (10x10).  For now, 0s will represent dead cells and 1s will represent live ones.  The Board class will implement the rules of the game as well as display the world, updating it each step.