Board now a 2-D array of Cell objects
Instead of merely being a 2-D array of 0s, the world is now represented internally as a 2-D array of Cell objects.  Note that as each Cell is added to the world, it is added with its x and y coordinates and that the Board class' print_world() method now uses each Cell's state field to display the current board configuration.