Cell class now obeys the rules of the game
The Cell class now has a check_rules method which counts its number of living neighbors and then may or may not change its own state depending on the rules of Conway's Game of Life and that number.  Also did a bit of cleaning up and organization in the Cell class.