Simulation now works as it is intended
The Board and Cell classes have been updated so that one can manually set a seed for the world and watch it for as many iterations as they would like.  Next goals are to 1) Make it look better, ideally running in a separate window with some minor graphics and 2) Make it run faster as it is currently not incredibly efficient.