GUI now updates correctly every cycle
I now have the GUI correctly updating itself every cycle of the simulation so you can actually watch Conway's Game of Life play out.  The only issue right now is that the window can no longer be closed via the 'x' button on the top of the window (presumably because I am calling the window's update method every frame but I am not sure) and so closing the program requires a Ctrl-c at the moment.
After fixing this I want to add in more functionality such as letting the user pick the starting seed as well as pause and resume at any time.