Added a bit more functionality and color
When ran, Main will now prompt the user for what dimensions they want the world to be and it will then generate a Board with those dimensions.  The Board class now also randomly selects a bunch of Cells on the Board to be alive when the simulation starts.  Finally, the Cells will change color based on how many cycles they have been alive thus far: cyan = 1 cycle, magenta = 2-4, green = 5-19 and black = 20 or more.  Also note that dead cells are gray now.  Still has to be quite via the command line but that fix is next.