Terminates when simulation has halted
The Board class now checks to see if the simulation has stagnated for a certain number of cycles in a row, terminating the simulation in such an instance.  Also, the Board now times how long the simulation has run, printing that run time to the console in the event of the simulation being terminated by the aforementioned conditions.